Improve stack traces on V8 so that they include the stack from...
...the call to the lifted function.  The stack isn't included normally, because
lifted functions are invoked by the observable mechanism.  Fortunately V8 shows
the names of function objects in stack traces.  So we etract the stack and give
it as the name to a wrapper function.  This has the downside that extracting the
name is likely to be somewhat expensive.  Another possibility would be to use a
`try-catch` wrapper, but that is not free either.